feat: :recycle: reorganizing folder structure
moving components to differents folders, separating components and logic
