+++ src/App.tsx
@@ -1,31 +1,16 @@
 import "./App.css";
-import { CharacterCard } from "./components/CharacterCard";
-import { Header } from "./components/Header";
-import { SearchChar } from "./components/SearchChar";
-import { Char, useChars } from "./hooks/useChars";
+import { Characters } from "./components/characters/Characters";
+import { Header } from "./components/header/Header";
+import { SearchChar } from "./components/searchInput/SearchChar";
+import { Title } from "./components/title/Title";
 
 function App() {
-  const { chars } = useChars();
   return (
     <div className="layout">
       <Header />
       <SearchChar />
-      <article className="layout__title">
-        <h1>lista de personajes</h1>
-        <hr />
-      </article>
-      <main className="layout__cardsContainer">
-        {chars.map((character: Char) => (
-          <CharacterCard
-            key={character.id}
-            name={character.name}
-            status={character.status}
-            species={character.species}
-            origin={character.origin}
-            image={character.image}
-          />
-        ))}
-      </main>
+      <Title />
+      <Characters />
     </div>
   );
 }
