+++ src/components/card/CharacterCard.tsx
