+++ src/components/characters/Characters.css
@@ -1,0 +1,25 @@
+.layout__cardsContainer {
+  width: 100%;
+  display: grid;
+  grid-template-rows: repeat(auto-fit, 1fr);
+  grid-template-columns: repeat(4, 1fr);
+  gap: 40px;
+}
+
+@media (width < 1360px) {
+  .layout__cardsContainer {
+    grid-template-columns: repeat(3, 1fr);
+  }
+}
+
+@media (width < 935px) {
+  .layout__cardsContainer {
+    grid-template-columns: repeat(2, 1fr);
+  }
+}
+
+@media (width <= 680px) {
+  .layout__cardsContainer {
+    grid-template-columns: repeat(1, 1fr);
+  }
+}
