+++ src/components/header/Header.tsx
@@ -1,5 +1,5 @@
-import logo from "../assets/Rick_and_Morty.svg.png";
-import logoMobile from "../assets/mobileHeaderLogo.png";
+import logo from "../../assets/Rick_and_Morty.svg.png";
+import logoMobile from "../../assets/mobileHeaderLogo.png";
 import "./Header.css";
 
 export const Header = () => {
