+++ src/components/searchInput/SearchChar.css
