+++ src/components/searchInput/SearchChar.tsx
