+++ src/components/title/Title.tsx
@@ -1,0 +1,10 @@
+import "./Title.css";
+
+export const Title = () => {
+  return (
+    <article className="layout__title">
+      <h1>lista de personajes</h1>
+      <hr />
+    </article>
+  );
+};
